Array/0035-search-insert-position/solution.go
/*
 * Copyright (c) 2020.  beihai@wingsxdu.com. All right reserved
 * ProjectName: $LeetCode
 * LastModified: 2/17/20, 10:16 PM
 * Author: beihai
 */

package _035_search_insert_position

func searchInsert(nums []int, target int) int {
	low, high := 0, len(nums)-1
	for low <= high {
		mid := low + (high-low)>>1
		if nums[mid] >= target {
			high = mid - 1
		} else if (mid == len(nums)-1) || (nums[mid+1] >= target) {
			return mid + 1
		} else {
			low = mid + 1
		}
	}
	return 0
}

func searchInsert1(nums []int, target int) int {
	low, high := 0, len(nums)-1
	for low <= high {
		mid := low + (high-low)>>1
		if nums[mid] >= target {
			high = mid - 1
		} else {
			if (mid == len(nums)-1) || (nums[mid+1] >= target) {
				return mid + 1
			}
			low = mid + 1
		}
	}
	return 0
}
